♻️ 불필요한 컴포넌트 구분을 제거하고, use client의 범위를 버튼 컴포넌트들로 좁힘
